Extract common macro code to `CommonMacroOps` (#835)
