+++ schema/shared/src/main/scala-2/zio/blocks/schema/CommonMacroOps.scala
@@ -1,5 +1,6 @@
 package zio.blocks.schema
 
+import zio.blocks.schema.{TypeName => SchemaTypeName}
 import scala.collection.mutable
 import scala.reflect.macros.blackbox
 import scala.reflect.NameTransformer
@@ -8,6 +9,23 @@ private[schema] object CommonMacroOps {
   def fail(c: blackbox.Context)(msg: String): Nothing = c.abort(c.enclosingPosition, msg)
 
   def typeArgs(c: blackbox.Context)(tpe: c.Type): List[c.Type] = tpe.typeArgs.map(_.dealias)
+
+  def companion(c: blackbox.Context)(tpe: c.Type): c.Symbol = {
+    import c.universe._
+
+    val comp = tpe.typeSymbol.companion
+    if (comp.isModule) comp
+    else {
+      val ownerChainOf = (s: Symbol) => Iterator.iterate(s)(_.owner).takeWhile(_ != NoSymbol).toArray.reverseIterator
+      val path         = ownerChainOf(tpe.typeSymbol)
+        .zipAll(ownerChainOf(c.internal.enclosingOwner), NoSymbol, NoSymbol)
+        .dropWhile(x => x._1 == x._2)
+        .takeWhile(x => x._1 != NoSymbol)
+        .map(x => x._1.name.toTermName)
+      if (path.isEmpty) NoSymbol
+      else c.typecheck(path.foldLeft[Tree](Ident(path.next()))(Select(_, _)), silent = true).symbol
+    }
+  }
 
   def directSubTypes(c: blackbox.Context)(tpe: c.Type): List[c.Type] = {
     import c.universe._
@@ -86,5 +104,59 @@ private[schema] object CommonMacroOps {
         subTypes.addOne(classType)
       }
     subTypes.toList
+  }
+
+  def typeName(
+    c: blackbox.Context
+  )(typeNameCache: mutable.HashMap[c.Type, SchemaTypeName[?]], tpe: c.Type): SchemaTypeName[?] = {
+    import c.universe._
+
+    def calculateTypeName(tpe: Type): SchemaTypeName[?] =
+      if (tpe =:= typeOf[java.lang.String]) SchemaTypeName.string
+      else {
+        var packages  = List.empty[String]
+        var values    = List.empty[String]
+        val tpeSymbol = tpe.typeSymbol
+        var name      = NameTransformer.decode(tpeSymbol.name.toString)
+        val comp      = companion(c)(tpe)
+        var owner     =
+          if (comp == null) tpeSymbol
+          else if (comp == NoSymbol) {
+            name += ".type"
+            tpeSymbol.asClass.module
+          } else comp
+        while ({
+          owner = owner.owner
+          owner.owner != NoSymbol
+        }) {
+          val ownerName = NameTransformer.decode(owner.name.toString)
+          if (owner.isPackage || owner.isPackageClass) packages = ownerName :: packages
+          else values = ownerName :: values
+        }
+        val tpeArgs = typeArgs(c)(tpe).map(ta => typeName(c)(typeNameCache, ta))
+        new SchemaTypeName(new Namespace(packages, values), name, tpeArgs)
+      }
+
+    typeNameCache.getOrElseUpdate(
+      tpe,
+      tpe match {
+        case TypeRef(compTpe, typeSym, Nil) if typeSym.name.toString == "Type" =>
+          var tpeName = calculateTypeName(compTpe)
+          if (tpeName.name.endsWith(".type")) tpeName = tpeName.copy(name = tpeName.name.stripSuffix(".type"))
+          tpeName
+        case _ =>
+          calculateTypeName(tpe)
+      }
+    )
+  }
+
+  def toTree(c: blackbox.Context)(tpeName: SchemaTypeName[?]): c.Tree = {
+    import c.universe._
+
+    val packages = tpeName.namespace.packages.toList
+    val values   = tpeName.namespace.values.toList
+    val name     = tpeName.name
+    val params   = tpeName.params.map(toTree(c)).toList
+    q"new TypeName(new Namespace($packages, $values), $name, $params)"
   }
 }
