+++ schema/shared/src/main/scala-2/zio/blocks/schema/SchemaCompanionVersionSpecific.scala
@@ -30,7 +30,6 @@ private object SchemaCompanionVersionSpecific {
 
   def derived[A: c.WeakTypeTag](c: blackbox.Context): c.Expr[Schema[A]] = {
     import c.universe._
-    import c.internal._
 
     def fail(msg: String): Nothing = CommonMacroOps.fail(c)(msg)
 
@@ -77,20 +76,7 @@ private object SchemaCompanionVersionSpecific {
       if (isZioPreludeNewtype(tpe)) zioPreludeNewtypeDealias(tpe)
       else tpe
 
-    def companion(tpe: Type): Symbol = {
-      val comp = tpe.typeSymbol.companion
-      if (comp.isModule) comp
-      else {
-        val ownerChainOf = (s: Symbol) => Iterator.iterate(s)(_.owner).takeWhile(_ != NoSymbol).toArray.reverseIterator
-        val path         = ownerChainOf(tpe.typeSymbol)
-          .zipAll(ownerChainOf(enclosingOwner), NoSymbol, NoSymbol)
-          .dropWhile(x => x._1 == x._2)
-          .takeWhile(x => x._1 != NoSymbol)
-          .map(x => x._1.name.toTermName)
-        if (path.isEmpty) NoSymbol
-        else c.typecheck(path.foldLeft[Tree](Ident(path.next()))(Select(_, _)), silent = true).symbol
-      }
-    }
+    def companion(tpe: Type): Symbol = CommonMacroOps.companion(c)(tpe)
 
     def primaryConstructor(tpe: Type): MethodSymbol = tpe.decls.collectFirst {
       case m: MethodSymbol if m.isPrimaryConstructor => m
@@ -129,52 +115,9 @@ private object SchemaCompanionVersionSpecific {
 
     val typeNameCache = new mutable.HashMap[Type, SchemaTypeName[?]]
 
-    def typeName(tpe: Type): SchemaTypeName[?] = {
-      def calculateTypeName(tpe: Type): SchemaTypeName[?] =
-        if (tpe =:= typeOf[java.lang.String]) SchemaTypeName.string
-        else {
-          var packages  = List.empty[String]
-          var values    = List.empty[String]
-          val tpeSymbol = tpe.typeSymbol
-          var name      = NameTransformer.decode(tpeSymbol.name.toString)
-          val comp      = companion(tpe)
-          var owner     =
-            if (comp == null) tpeSymbol
-            else if (comp == NoSymbol) {
-              name += ".type"
-              tpeSymbol.asClass.module
-            } else comp
-          while ({
-            owner = owner.owner
-            owner.owner != NoSymbol
-          }) {
-            val ownerName = NameTransformer.decode(owner.name.toString)
-            if (owner.isPackage || owner.isPackageClass) packages = ownerName :: packages
-            else values = ownerName :: values
-          }
-          new SchemaTypeName(new Namespace(packages, values), name, typeArgs(tpe).map(typeName))
-        }
+    def typeName(tpe: Type): SchemaTypeName[?] = CommonMacroOps.typeName(c)(typeNameCache, tpe)
 
-      typeNameCache.getOrElseUpdate(
-        tpe,
-        tpe match {
-          case TypeRef(compTpe, typeSym, Nil) if typeSym.name.toString == "Type" =>
-            var tpeName = calculateTypeName(compTpe)
-            if (tpeName.name.endsWith(".type")) tpeName = tpeName.copy(name = tpeName.name.stripSuffix(".type"))
-            tpeName
-          case _ =>
-            calculateTypeName(tpe)
-        }
-      )
-    }
-
-    def toTree(tpeName: SchemaTypeName[?]): Tree = {
-      val packages = tpeName.namespace.packages.toList
-      val values   = tpeName.namespace.values.toList
-      val name     = tpeName.name
-      val params   = tpeName.params.map(toTree).toList
-      q"new TypeName(new Namespace($packages, $values), $name, $params)"
-    }
+    def toTree(tpeName: SchemaTypeName[?]): Tree = CommonMacroOps.toTree(c)(tpeName)
 
     def modifiers(tpe: Type): List[Tree] = {
       val modifiers = new mutable.ListBuffer[Tree]
